Introduce configurable throttle
This commit allows generators to provide a configuration option for their
throttle, here only udp is converted. The ultimate ambition is to allow for two
additional throttle variants: all-out and stable. We rename the existing
throttle to 'predictive' and set this as the default. Note that the throttle is
implemented in terms of an enum. I did attempt to make this static dispatch
because throttle is called in our hot loops but found the intrusion into the
codebase too high, especially with consideration of the throttle being present
determined at runtime.

REF SMP-520
REF SMP-521
