+++ src/generator/udp.rs
@@ -17,7 +17,7 @@ use crate::{
     block::{self, chunk_bytes, construct_block_cache, Block},
     payload,
     signals::Shutdown,
-    throttle::Throttle,
+    throttle::{self, Throttle},
 };
 
 #[derive(Debug, Deserialize, PartialEq)]
@@ -35,6 +35,9 @@ pub struct Config {
     pub block_sizes: Option<Vec<byte_unit::Byte>>,
     /// The maximum size in bytes of the cache of prebuilt messages
     pub maximum_prebuild_cache_size_bytes: byte_unit::Byte,
+    /// The load throttle configuration
+    #[serde(default)]
+    pub throttle: throttle::Config,
 }
 
 /// Errors produced by [`Udp`].
@@ -118,10 +121,11 @@ impl Udp {
             .expect("could not convert to socket")
             .next()
             .unwrap();
+
         Ok(Self {
             addr,
             block_cache,
-            throttle: Throttle::new(bytes_per_second),
+            throttle: Throttle::new_with_config(config.throttle, bytes_per_second),
             metric_labels: labels,
             shutdown,
         })
